+++ R/power_scABEL.R
@@ -6,7 +6,7 @@
 #---------------------------------------------------------------------------
 power.scABEL <- function(alpha=0.05, theta1, theta2, theta0, CV, n,   
                          design=c("2x3x3", "2x2x4", "2x2x3"), regulator,
-                         nsims=1E5, details=FALSE, setseed=TRUE)
+                         nsims, details=FALSE, setseed=TRUE)
 {
   # design must be checked outside
   desi <- match.arg(design)