+++ R/power_scABEL2.R
@@ -15,7 +15,7 @@
 
 power.scABEL2 <- function(alpha=0.05, theta1, theta2, theta0, CV, n,   
                           design=c("2x3x3", "2x2x4", "2x2x3"), regulator,
-                          nsims=1E5, details=FALSE, setseed=TRUE)
+                          nsims, details=FALSE, setseed=TRUE)
 {
   # function is no longer visible via NAMESPACE
   # .Deprecated(new="power.scABEL", msg=paste0("'power.scABEL2' is deprecated.",
@@ -36,6 +36,15 @@ power.scABEL2 <- function(alpha=0.05, theta1, theta2, theta0, CV, n,
   
   CVwT <- CV[1]
   if (length(CV)==2) CVwR <- CV[2] else CVwR <- CVwT
+
+  if (missing(nsims)) { # not given
+    if (theta0 == scABEL(CVwR)[["lower"]] | theta0 == scABEL(CVwR)[["upper"]]) {
+      nsims <- 1e6 # simulating TIE
+    } else {
+      nsims <- 1e5 # simulating power
+    }
+  }
+  
   s2wT <- log(1.0 + CVwT^2)
   s2wR <- log(1.0 + CVwR^2)